Move autocompletion helper to separate module
