+++ watson/utils.py
@@ -17,6 +17,7 @@ except ImportError:
 import click
 import arrow
 
+import watson as _watson
 from .fullmoon import get_last_full_moon
 
 from click.exceptions import UsageError
@@ -415,3 +416,7 @@ def json_arrow_encoder(obj):
         return obj.for_json()
 
     raise TypeError("Object {} is not JSON serializable".format(obj))
+
+
+def get_watson_instance():
+    return _watson.Watson(config_dir=os.environ.get('WATSON_DIR'))
